% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-unwarp.R
\name{formula.warped_model}
\alias{formula.warped_model}
\title{Formula method for warped models}
\usage{
\method{formula}{warped_model}(x, warped = TRUE, ...)
}
\arguments{
\item{x}{Warped model, of class \code{warped_model}.}

\item{warped}{Return the formula involving the warped predictors
or the original untransformed predictor names. Defaults to \code{TRUE},
i.e. warped names, since the warped model wants to be talked to
in the language of warped predictors.}

\item{...}{Not used, ignored with a warning.}
}
\value{
Formula
}
\description{
Formula method for warped models
}
