% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gabor}
\alias{Gabor}
\title{Rock glacier remote sensing using Gabor texture filters}
\format{
A data frame with 3403 rows and 49 variables.
\describe{
\item{class}{Factor variable indicating that the observation represents
flow structures on a rock glacier (\code{"TRUE"}) or not
(\code{"FALSE"}).}
\item{...}{Terrain attributes and Gabor texture features as predictors.}
}
}
\source{
Brenning, A., Long, S. and Fieguth, P. 2012. Detecting rock glacier
flow structures using Gabor filters and IKONOS imagery. Remote Sensing
of Environment, 125: 227-237. https://doi.org/10.1016/j.rse.2012.07.005
}
\usage{
Gabor
}
\description{
This is a subset of the data used by Brenning et al. (2012).
}
\keyword{datasets}
