% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbound-outbound.R
\name{inbound.pca_warper}
\alias{inbound.pca_warper}
\title{Methods for forward (outbound) and backward (inbound) feature space transformation using warper objects}
\usage{
\method{inbound}{pca_warper}(object, wdata = object$pca$x, ...)
}
\arguments{
\item{object}{A \code{warper} object, e.g. created by calling \code{\link[=pca_warper]{pca_warper()}}
for a principal component transformation.}

\item{wdata}{A data frame containing data in transformed space, i.e.
coordinates with respect to principal axes.}

\item{...}{Additional arguments to be passed to \code{\link[=inbound.rotation_warper]{inbound.rotation_warper()}}}
}
\value{
A transformed data frame of class \code{warped_df} (for outbound)
or \code{data.frame} (for inbound transformations).
}
\description{
Outbound or forward transformation refers to the mapping \eqn{T}
from feature space into a transformed space (e.g., principal components space),
and inbound or backward transformation is its inverse transformation \eqn{T^{1}}
back into feature space.
}
\details{
These functions are mainly intended for internal use.
The \code{\link[=warp]{warp()}} and \code{\link[=unwarp]{unwarp()}} methods are more
user-oriented interfaces.
}
