% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Maipo}
\alias{Maipo}
\title{Maipo fruit-tree crop data}
\format{
A data frame with 400 rows and 67 variables:
\describe{
\item{class}{Croptype \code{"crop1"} to \code{"crop4"} as a factor variable.}
\item{utmx,utmy}{UTM x and y coordinates}
\item{b12,...,b87}{Landsat reflectances: variable \code{b}ij corresponds to
band j from image date i, where larger values of i indicate a date later
in the season. See referenced paper for details.}
\item{ndvi01,...,ndvi08}{NDVI values for each image date.}
\item{ndwi01,...,ndwi}{NDWI values, i.e. a moisture index.}
}
}
\source{
Peña, M.A., and Brenning, A. 2015. Assessing Fruit-Tree Crop
Classification from Landsat-8 Time Series for the Maipo Valley, Chile.
Remote Sensing of Environment 171: 234-44. \url{https://doi.org/10.1016/j.rse.2015.10.029}
}
\usage{
Maipo
}
\description{
Maipo data - a subset of the data used by Pena and Brenning (2015), aggregated
to the field level, i.e. each field or parcel is one observation.
}
\keyword{datasets}
