% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inbound-outbound.R
\name{outbound}
\alias{outbound}
\alias{inbound}
\title{Generic methods for forward (outbound) and backward (inbound) feature space transformation using warper objects}
\usage{
outbound(object, ...)

inbound(object, ...)
}
\arguments{
\item{object}{A \code{warper} object, e.g. created by calling \code{\link[=pca_warper]{pca_warper()}}
for a principal component transformation.}

\item{...}{Additional arguments for the corresponding method.}
}
\value{
A transformed data frame of class \code{warped_df} (for outbound)
or \code{data.frame} (for inbound transformations).
}
\description{
Outbound or forward transformation refers to the mapping \eqn{T}
from feature space into a transformed space (e.g., principal component space),
and inbound or backward transformation is its inverse transformation \eqn{T^{1}}
back into feature space.
}
\details{
These functions are mainly intended for internal use.
The \code{\link[=warp]{warp()}} and \code{\link[=unwarp]{unwarp()}} methods are more
user-oriented interfaces.
}
\section{Functions}{
\itemize{
\item \code{inbound}: Generic method for inbound
}}

