% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{pca_warper}
\alias{pca_warper}
\title{Principal components transformation of feature space}
\usage{
pca_warper(xdata, xvars, wvars = "PC", yvar, uvars = NULL, title = wvars)
}
\arguments{
\item{xdata}{A data frame containing the observations in the original feature space.}

\item{xvars}{A character vector with the column names of features in \code{xdata}
that should be transformed.}

\item{wvars}{A character vector with the names of all transformed predictors,
or (more often) a simple character string to which the numbers 1, ...
will be appended as needed.}

\item{yvar}{Name of the response variable (not to be transformed)}

\item{uvars}{Names of additional variables that should remain untouched.}

\item{title}{Optional name of the transformation, may be used for printing summaries
or for plotting.}
}
\value{
An object of class \code{warper}, \code{rotation_warper} and \code{pca_warper}.
}
\description{
This function generates a \code{warper} object based on a principal components
analysis of the features (or more precisely, of selected features \code{xvars}).
}
\details{
The function will make an effort to ensure that there's no overlap
between \code{xvars}, \code{uvars} and \code{yvar}; however, ideally
these arguments should not overlap.
}
