% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{plot.pca_warper}
\alias{plot.pca_warper}
\title{Plot a PCA warper transformation object}
\usage{
\method{plot}{pca_warper}(x, which = c(1:2), col = c("white", "black"), bp.col = "lightblue", ...)
}
\arguments{
\item{which}{Which plots to plot: screeplot (\code{1}), biplot (\code{2}).}

\item{col, bp.col}{Colours of \code{biplot} labels and \code{barplot} bars.}

\item{...}{Currently not used.}

\item{object}{A \code{\link{pca_warper}} object.}
}
\value{
Returns \code{NULL}
}
\description{
This method plots standard PCA summary diagrams.
}
\examples{
### Create principal components warper for Maipo data set:
xvars <- c(paste("ndvi0", 1:8, sep = ""), paste("ndwi0", 1:8, sep = ""),
           paste("b", outer(1:8,2:7,paste,sep = ""), sep = ""))
fo <- as.formula(paste("class ~", paste(xvars, collapse=" +" )))
d <- maipo

wrp <- pca_warper(d, xvars = xvars, yvar = "class")
plot(wrp)
round(wrp$pca$rotation[,1:3], digits = 2)
summary(wrp$pca)

# Create a data frame with transformed data:
wd <- warp(d, warper = wrp)
summary(wd)

# Backtransform it, should be identical to d:
d2 <- unwarp(wd, warper = wrp)
all.equal(d, d2[,colnames(d)])
# Default tolerance works for this data set, but you may have to use
# e.g. tol = 10^(-6) for less well conditioned data sets and
# transformations.
}
