% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{predict.warped_model}
\alias{predict.warped_model}
\title{Predict from a warped fitted machine-learning model}
\usage{
\method{predict}{warped_model}(object, newdata = NULL, force_unwarp = TRUE, ...)
}
\arguments{
\item{object}{A \code{warped_model} object, i.e. a warped fitted machine-learning model as created by \code{\link{warp_fitted_model}}.}

\item{newdata}{A data frame with features in transformed space. This can be a \code{warped_df} object or just a plain \code{data.frame}. See \code{force_unwarp}.}

\item{force_unwarp}{If \code{TRUE}, \code{newdata} will always be (back)transformed
into feature space using the \code{warper}'s inverse transformation.
If \code{FALSE}, only objects of class \code{warped_df} will be backtransformed.}

\item{...}{Additional arguments for the predict method of \code{object$fit}.}
}
\value{
The function's result depends on the behaviour of the invoked
predict method.
}
\description{
This function implements the prediction \eqn{\textbf{w}\mapsto\hat{y}:=(\hat{f}\circ T^{-1})(\textbf{w})}
by (back)transforming a feature vector \eqn{w} from the transformed space (e.g.,
principal components space) to the original feature space and then applying
the fitted machine learning model's predict method to make predictions.
}
\details{
\code{newdata} will normally be a data frame, not \code{NULL}; not all
fitted models accept \code{NULL}. The most common behaviour is to return
model predictions for the training data.
}
\seealso{
\code{\link{warp_fitted_model}} for creating a warped machine-learning
model, and the \code{\link{predict}} method of your machine-learning model
for details on additional arguments and the returned values.
}
