% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_warper.R
\name{print.summary.pca_warper}
\alias{print.summary.pca_warper}
\title{Print summary of \code{pca_warper} object}
\usage{
\method{print}{summary.pca_warper}(x, ...)
}
\arguments{
\item{x}{\code{summary.pca_warper} object for principal component transformation of feature space}

\item{...}{Additional arguments for \code{\link[stats:prcomp]{stats::print.summary.prcomp()}},}
}
\value{
Returns the \code{summary.pca_warper} object.
}
\description{
Print summary of \code{pca_warper} object
}
\examples{
xvars <- c(paste("ndvi0", 1:8, sep = ""),
           paste("ndwi0", 1:8, sep = ""))
fo <- as.formula(paste("class ~", paste(xvars, collapse=" +" )))
wrp <- pca_warper(maipo, xvars = xvars, yvar = "class")
summary(wrp)
}
\seealso{
\code{\link[=summary.pca_warper]{summary.pca_warper()}}, \code{\link[stats:prcomp]{stats::print.summary.prcomp()}}
}
