% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strucpca_warper.R
\name{strucpca_warper}
\alias{strucpca_warper}
\title{Structured principal components transformation of feature space}
\usage{
strucpca_warper(
  xdata,
  xvars,
  wvars = NULL,
  yvar,
  uvars = NULL,
  center = TRUE,
  scale = TRUE,
  positive = TRUE,
  title = NULL
)
}
\arguments{
\item{xdata}{A data frame containing the observations in the original feature space.}

\item{xvars}{A list of character vectors with the column names of features in \code{xdata}
that should be transformed. Variables named in each list component are jointly
transformed using PCA. These groups shouldn't be too correlated amongst
each other.}

\item{wvars}{A character vector of same length as \code{xvars} giving the prefixes
to be used to generate the names of transformed predictors. If of length 1,
use prefix \code{"<wvars>[1]"} for first feature subset, i.e. \code{"PC[1]2"} for
second principal component from first feature subset. If \code{NULL}, use
\code{"PC"} as the prefix.}

\item{yvar}{Name of the response variable (not to be transformed)}

\item{uvars}{Optional list of same length as \code{xvars} with names of additional
variables that should remain untouched.}

\item{center}{Logical arguments indicating whether the data should
be centered and then scales. Both should be turned on (\code{TRUE}), which is
the default behaviour.}

\item{scale}{Logical arguments indicating whether the data should
be centered and then scales. Both should be turned on (\code{TRUE}), which is
the default behaviour.}

\item{positive}{Logical argument (default: \code{TRUE}) indicating whether
the signs of the loadings should be adjusted so that the most
strongly weighted PC gets a positive sign.}

\item{title}{Optional name of the transformation, may be used for printing summaries
or for plotting.}
}
\value{
An object of class \code{warper}, \code{rotation_warper} and \code{strucpca_warper}.
}
\description{
This function offers a more structured approach to feature space transformation
by allowing the user to transform different groups of predictor variables
separately. It generates a \code{warper} object based on a principal components
analyes applied to the feature subsets.
}
