% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_warper.R
\name{summary.pca_warper}
\alias{summary.pca_warper}
\title{Summary of a \code{pca_warper} object}
\usage{
\method{summary}{pca_warper}(object, ...)
}
\arguments{
\item{object}{A \code{pca_warper} object.}

\item{...}{Additional arguments to \code{\link[stats:prcomp]{stats::summary.prcomp()}}
(currently not used).}
}
\value{
A \code{summary.pca_warper} object.
}
\description{
Summary of a \code{pca_warper} object
}
\examples{
xvars <- c(paste("ndvi0", 1:8, sep = ""),
           paste("ndwi0", 1:8, sep = ""))
fo <- as.formula(paste("class ~", paste(xvars, collapse=" +" )))
wrp <- pca_warper(maipo, xvars = xvars, yvar = "class")
summary(wrp)
}
\seealso{
\code{\link[stats:prcomp]{stats::summary.prcomp()}}
}
