% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{unwarp}
\alias{unwarp}
\title{Backtransform from transformed ('warped') to original feature space}
\usage{
unwarp(x, ...)
}
\arguments{
\item{x}{The object to be backtransformed, for example a data frame or a \code{warped_df} object.}

\item{...}{Additional arguments to be passed to the \code{unwarp} method.}
}
\value{
A data frame with features in the original feature space.
}
\description{
These methods apply the inverse transformation \eqn{T^{-1}} to data
in transformed space - e.g. a feature vector in principal components
space back to the original feature space, in the case of a PCA
transformation.
}
\seealso{
\code{\link{warp}} for the forward transformation, and
\code{\link{pca_warper}} for an example of a simple warper function based
on the principal components transformation.
}
