% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{unwarp.warped_df}
\alias{unwarp.warped_df}
\alias{unwarp.data.frame}
\title{Backtransform from transformed ('warped') to original feature space}
\usage{
\method{unwarp}{warped_df}(x, warper, force_unwarp = FALSE)

\method{unwarp}{data.frame}(x, warper = NULL, force_unwarp = FALSE)
}
\arguments{
\item{x}{The object to be backtransformed, for example a data frame or a \code{warped_df} object.}

\item{warper}{A \code{warper} object as, for example, created by \code{\link{pca_warper}} for principal components transformations.}

\item{force_unwarp}{If \code{TRUE}, backtransform even if it's not a \code{warped_df} object. (Ignored by \code{unwarp.warped_df}.)}

\item{...}{Additional arguments to be passed to the \code{unwarp} method.}
}
\value{
A data frame with features in the original feature space.
}
\description{
These methods apply the inverse transformation \eqn{T^{-1}} to data
in transformed space - e.g. a feature vector in principal components
space back to the original feature space, in the case of a PCA
transformation.
}
\section{Functions}{
\itemize{
\item \code{unwarp.data.frame}: Backtransform a data frame from transformed to original feature space
}}

\seealso{
\code{\link{warp}} for the forward transformation, and
\code{\link{pca_warper}} for an example of a simple warper function based
on the principal components transformation.
}
