% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-unwarp.R
\name{warp.formula}
\alias{warp.formula}
\alias{unwarp.formula}
\title{Warp model formula}
\usage{
\method{warp}{formula}(x, warper, ...)

\method{unwarp}{formula}(x, warper, ...)
}
\arguments{
\item{x}{Formula object.}

\item{warper}{Warper object.}

\item{...}{Not used, ignored with a warning.}
}
\value{
The warped formula (e.g., based on principal components), or the
unwarped formula (using the original, untransformed variable names).
}
\description{
Warp model formula
}
\details{
Currently only "simple" formulas of the form \code{y ~ x1 + x2 + x3} etc.
are allowed - no \code{*}, \code{:}, \code{s()}, \code{I()}, \verb{^2} or other operations that
are supported by some modeling functions.
}
\section{Functions}{
\itemize{
\item \code{unwarp.formula}: Unwarp a model formula.
}}

