% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{warp.function}
\alias{warp.function}
\title{Create a warped version of a model fitting function}
\usage{
\method{warp}{`function`}(x, warper, ...)
}
\arguments{
\item{x}{A model fitting function such as \code{\link{lm}} or \code{\link[randomForest]{randomForest}}}

\item{warper}{A \code{warper} object, for example a \code{\link{pca_warper}}.}

\item{...}{Additional arguments to be passed to the model fitting function.}
}
\value{
A function representing the composition of the learning function \eqn{F:L\mapsto\hat{f}}
with the inverse transformation function \eqn{T^{-1}}. In other words,
this function first (back)transformed its input data from the transformed space
to the original feature space, and then fits a model using these features.
}
\description{
Use \code{\link{warp_fitted_model}} instead, which provides a more
elegant way to create a warped perspective on a model. This function may
be omitted in the future.
}
\seealso{
\code{\link{warp_fitted_model}}
}
