% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-unwarp.R
\name{warp}
\alias{warp}
\title{Transform data from feature space into a transformed space}
\usage{
warp(x, ...)
}
\arguments{
\item{x}{The object to be transformed, normally a data frame.}

\item{...}{Additional arguments to be passed to the \code{warp} method.}
}
\value{
A data frame with features in the transformed (e.g., PCA) space.
}
\description{
These methods apply the transformation \eqn{T} to data
in feature space - the result is, for example, the principal components
representation of the data, if a \link{pca_warper} is used.
}
\seealso{
\code{\link[=unwarp]{unwarp()}} for the inverse transformation, and
\code{\link[=pca_warper]{pca_warper()}} for an example of a simple warper function based
on the principal component transformation.
}
