% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{warp_fitted_model}
\alias{warp_fitted_model}
\title{Create a warped view of a fitted machine-learning model}
\usage{
warp_fitted_model(x, warper)
}
\arguments{
\item{x}{A fitted model, for example from a \code{randomForest} call.
This model was fitted to the original, untransformed data.}

\item{warper}{A \code{warper} object, representing an (invertible)
transformation from the model's feature space to a transformed space
whose axes (or base vectors) define a new perspective on the
feature space and model.}
}
\value{
An object of class \code{warped_model}, which can be used to
make predictions using input data from the transformed space
thanks to the \code{predict} method for this object type.
}
\description{
This function creates the composition function \eqn{\hat{f}\circ T^{-1}}
from a fitted machine-learning model \eqn{\hat{f}} and a warper object
representing the transformation function \eqn{T}.
}
\details{
This function does not re-fit the model, it simply creates a
data structure that jointly stores the fitted model and the warper
object. The \code{warped_model}'s \code{predict} (and other) methods
allow the user to use a \code{warped_model} as if it had been fitted
using the transformed data.
}
