% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gabor}
\alias{gabor}
\title{Rock glacier remote sensing using Gabor texture filters}
\format{
A data frame with 4656 grid cells (1290
from flow patterns and 3366 from other terrain outside of
rock glaciers) in two study areas (\code{area = "LN"}: 2499 grid cells;
\code{area = "CAT"}: 2157 cells).
(The flow patterns are from approximately 50 individual rock glaciers
of varying size.) \describe{
\item{class}{Factor variable (levels: \code{"TRUE"}, \code{"FALSE"}) representing the
presence (\code{"TRUE"}) and absence (\code{"FALSE"}) of rock glacier flow patterns.}
\item{x,y}{UTM x and y coordinates (not to be used as predictors)}
\item{area}{Factor variable identifying one of the two study
areas, \code{"LN"} or \code{"CAT"}.}
\item{dem}{Elevation in metres above sea level (m a.s.l.)}
\item{slope}{Local slope angle in degrees}
\item{cslope}{Slope angle of the upslope contributing area in degrees}
\item{log.carea}{Logarithm (to the base 10) of the size of the upslope
contributing area in m²}
\item{log.cheight}{Logarithm (to the base 10) of the height of the upslope contributing area in m}
\item{pisr}{Annual potential incoming solar radiation}
\item{m30e1g5rg,m30e1g5max,m30e1g5min,m30e1g5med,m30e1g10rg,m30e1g10max,m30e1g10min,m30e1g10med,m30e1g20rg,m30e1g20max,m30e1g20min,m30e1g20med,m30e1g30rg,m30e1g30max,m30e1g30min,m30e1g30med,m30e1g50rg,m30e1g50max,m30e1g50min,m30e1g50med,m30e2g5rg,m30e2g5max,m30e2g5min,m30e2g5med,m30e2g10rg,m30e2g10max,m30e2g10min,m30e2g10med,m30e2g20rg,m30e2g20max,m30e2g20min,m30e2g20med,m30e2g30rg,m30e2g30max,m30e2g30min,m30e2g30med,m30e2g50rg,m30e2g50max,m30e2g50min,m30e2g50med}{Gabor features of the form \code{m30e*i*g*j*x} with the following settings (see Brenning et al. 2012 for details): \emph{i} = axis ratio (1 or 2) of Gabor filter; \emph{j} = wavelength of Gabor filter (5, 10, 20, 30, or 50 m); \strong{x} = aggregation scheme (“min” = minimum; “max” = maximum; “rg” = range; “med” = median}}
}
\source{
Brenning, A., Long, S. and Fieguth, P. 2012. Detecting rock glacier
flow structures using Gabor filters and IKONOS imagery. \emph{Remote Sensing of
Environment}, 125: 227-237. \url{https://doi.org/10.1016/j.rse.2012.07.005}
}
\usage{
gabor
}
\description{
This data set for binary classification tasks contains information on the
presence/absence of flow structures related to the deformation of rock
glaciers in the Andes of Santiago, and corresponding remotely-sensed texture
attributes and terrain attributes as predictors. Rock glaciers are ice-debris
landforms - the visible expression of creeping mountain permafrost.
}
\details{
Texture attributes derived from high-resolution panchromatic IKONOS
imagery form the largest group of features in this study, and terrain
attributes are used as additional predictors. A ‘filter bank’ of Gabor
filters is used since Gabor features are capable of detecting ‘zebra
stripe’ type patterns that relate to the troughs and ridges typically found
on rock glaciers. The second group of features represent topographic
conditions, e.g. elevation, slope angle and size of the upslope
contributing area. These are proxies for topoclimatic conditions that may
relate to the formation and conservation of permafrost, and for talus
supply to these ice-debris landforms.

This data set is a subset of the data used by Brenning et al. (2012),
specifically a subset of the Laguna Negra area. Note that areas that can
“obviously” not present rock glaciers have been masked out (i.e. removed
from the data set), e.g. steep slopes, in order to allow the classifier to
focus on the “difficult” areas; see Brenning et al. (2012) for details.

\strong{Objective:} To identify rock-glacier flow patterns based on the available
texture and terrain attribute data.
}
\keyword{datasets}
