% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{inbound.pca_warper}
\alias{inbound.pca_warper}
\title{Methods for forward (outbound) and backward (inbound) feature space transformation using warper objects}
\usage{
\method{inbound}{pca_warper}(object, wdata = object$pca$x, ...)
}
\arguments{
\item{object}{A \code{warper} object created by \code{\link{pca_warper}}
for principal components transformations.}

\item{wdata}{A data frame containing data in transformed space, i.e.
coordinates with respect to principal components axes.}

\item{...}{Additional arguments to be passed to \code{inbound.rotation_warper}}
}
\value{
A transformed data frame of class \code{warped_df} (for outbound)
or \code{data.frame} (for inbound transformations).
}
\description{
Outbound or forward transformation refers to the mapping \eqn{T}
from feature space into a transformed space (e.g., principal components space),
and inbound or backward transformation is its inverse transformation \eqn{T^{1}}
back into feature space.
}
\details{
These functions are mainly intended for internal use.
The \code{\link{warp}} and \code{\link{unwarp}} methods are more
user-friendly.
}
