% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{inbound.rotation_warper}
\alias{inbound.rotation_warper}
\alias{outbound.rotation_warper}
\title{Methods for forward (outbound) and backward (inbound) feature space transformation using warper objects}
\usage{
\method{inbound}{rotation_warper}(object, wdata = object$wdata)

\method{outbound}{rotation_warper}(object, xdata = NULL)
}
\arguments{
\item{object}{A \code{warper} object.}

\item{wdata}{A data frame containing data in transformed space.}

\item{xdata}{A data frame with features in the original, untransformed
feature space.}
}
\value{
A transformed data frame of class \code{warped_df} (for outbound)
or \code{data.frame} (for inbound transformations).
}
\description{
Outbound or forward transformation refers to the mapping \eqn{T}
from feature space into a transformed space (e.g., principal components space),
and inbound or backward transformation is its inverse transformation \eqn{T^{1}}
back into feature space. This is an abstract parent class for \link{pca_warper}
and other related objects that can be expressed using rotation matrices.
}
\details{
These functions are mainly intended for internal use.
The \code{\link{warp}} and \code{\link{unwarp}} methods are more
user-friendly.
}
\section{Functions}{
\itemize{
\item \code{outbound.rotation_warper}: Forward (outbound) transformation using a rotation warper
}}

