% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{maipofields}
\alias{maipofields}
\title{Remote sensing of fruit trees}
\format{
A data frame with 400 rows and 67 variables:
\describe{
\item{class}{Croptype \code{"crop1"} to \code{"crop4"} as a factor
variable.}
\item{utmx,utmy}{UTM x and y coordinates - not to be used as
predictors.}
\item{b12,b13,b14,b15,b16,b17,b22,b23,b24,b25,b26,b27,b32,b33,b34,b35,b36,b37,b42,b43,b44,b45,b46,b47,b52,b53,b54,b55,b56,b57,b62,b63,b64,b65,b66,b67,b72,b73,b74,b75,b76,b77,b82,b83,b84,b85,b86,b87}{\code{b}\emph{ij}
is the value of Landsat band \emph{j} in image \emph{i}; images are
numbered from early season = 1 to late season = 9; see Peña et al. (2017)
for details, e.g. image dates.}
\item{ndvi01,ndvi02,ndvi03,ndvi04,ndvi05,ndvi06,ndvi07,ndvi08}{\code{ndvi0}\emph{i}
is the NDVI value (representing vegetation vigour) for
each image date \emph{i}.}
\item{ndwi01,ndwi02,ndwi03,ndwi04,ndwi05,ndwi06,ndwi07,ndwi08}{\code{ndwi0}\emph{i}
is the NDWI value for image date \emph{i}, i.e. a moisture index (representing moisture
content of plants).} }
}
\source{
Pena, M.A., and Brenning, A. 2015. Assessing fruit-tree crop
classification from Landsat-8 time series for the Maipo Valley, Chile.
\emph{Remote Sensing of Environment}, 171: 234-44.
\url{https://doi.org/10.1016/j.rse.2015.10.029}
}
\usage{
maipofields
}
\description{
This data set for multi-class classification contains inventoried fruit-tree
crop types and multitemporal Landsat satellite data from the Maipo river
basin in central Chile. And Landsat data is available for nine image dates
from one growing season. This is a subset of the data used by Pena et al.
(2015), who provide more detailed information on the data set (e.g., image
dates). Note that this is different from the \link[sperrorest:maipo]{sperrorest::maipo} data set shipped with
the \code{sperrorest} package, which contains pixel-level data; this data set
contains field-level data, i.e. one (averaged) observation per field (n=400).
}
\details{
\strong{Objective:} To predict crop type as the response variable based
on the available optical remote sensing data.
}
\keyword{datasets}
