% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{outbound}
\alias{outbound}
\alias{inbound}
\title{Generic methods for forward (outbound) and backward (inbound) feature space transformation using warper objects}
\usage{
outbound(object, ...)

inbound(object, ...)
}
\arguments{
\item{object}{A \code{warper} object, such as a \code{\link{pca_warper}}
for principal components transformations.}
}
\value{
A transformed data frame of class \code{warped_df} (for outbound)
or \code{data.frame} (for inbound transformations).
}
\description{
Outbound or forward transformation refers to the mapping \eqn{T}
from feature space into a transformed space (e.g., principal components space),
and inbound or backward transformation is its inverse transformation \eqn{T^{1}}
back into feature space.
}
\details{
These functions are mainly intended for internal use.
The \code{\link{warp}} and \code{\link{unwarp}} methods are more
user-friendly.
}
\section{Functions}{
\itemize{
\item \code{inbound}: Generic method for inbound
}}

