% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{plot.pca_warper}
\alias{plot.pca_warper}
\title{Plot of a PCA warper transformation object}
\usage{
\method{plot}{pca_warper}(object, col = c("white", "black"), bp.col = "lightblue")
}
\arguments{
\item{object}{A \link{pca_warper} object.}

\item{col, bp.col}{Colours of \code{biplot} labels and \code{barplot} bars.}
}
\value{
Invisibly returns the result of the biplot call (this may change)
}
\description{
This method plots standard PCA summary diagrams.
}
