% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{pls_warper}
\alias{pls_warper}
\title{Partial least squares transformation of feature space}
\usage{
pls_warper(
  xdata,
  xvars,
  pvars,
  wvars = "resid",
  yvar,
  uvars = NULL,
  title = wvars
)
}
\arguments{
\item{xdata}{A data frame containing the observations in the original feature space.}

\item{xvars}{A character vector with the column names of features in \code{xdata}
that should be transformed.}

\item{pvars}{A character string identifying one feature in \code{xdata}
that is used as the selected feature; the other features in
\code{xvars} will be transformed by orthogonalization, while
this one will remain unchanged.}

\item{wvars}{A character string giving a prefix for partial residual features.}

\item{yvar}{Name of the response variable (not to be transformed)}

\item{uvars}{Names of additional variables that should remain untouched.}

\item{title}{Optional name of the transformation, may be used for printing summaries
or for plotting.}
}
\value{
An object of class \code{warper}, \code{rotation_warper} and \code{pls_warper}.
}
\description{
This function generates a \code{warper} object based on a partial least squares
transformation with respect to a selected feature.
}
\details{
The arguments
\code{pvars}, \code{xvars}, \code{uvars} and \code{yvar} should not overlap.
}
