% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warp-unwarp.R
\name{unwarp.warped_df}
\alias{unwarp.warped_df}
\alias{unwarp.data.frame}
\title{Backtransform from transformed ('warped') to original feature space}
\usage{
\method{unwarp}{warped_df}(x, warper, force_unwarp = FALSE, ...)

\method{unwarp}{data.frame}(x, warper = NULL, force_unwarp = FALSE, ...)
}
\arguments{
\item{x}{The object to be backtransformed, for example a data frame or a \code{warped_df} object.}

\item{warper}{A \code{warper} object as, for example, created by \code{\link[=pca_warper]{pca_warper()}} for principal components transformations.}

\item{force_unwarp}{If \code{TRUE}, backtransform even if it's not a \code{warped_df} object. (Ignored by \code{unwarp.warped_df}.)}

\item{...}{Additional arguments to be passed to the \code{\link[=unwarp]{unwarp()}} method.}
}
\value{
A data frame with features in the original feature space.
}
\description{
These methods apply the inverse transformation \eqn{T^{-1}} to data
in transformed space - e.g. a feature vector in principal components
space back to the original feature space, in the case of a PCA
transformation.
}
\section{Functions}{
\itemize{
\item \code{unwarp.data.frame}: Backtransform a data frame from transformed to original feature space
}}

\examples{
### Create principal components warper for Maipo data set:
xvars <- c(paste("ndvi0", 1:8, sep = ""), paste("ndwi0", 1:8, sep = ""),
           paste("b", outer(1:8,2:7,paste,sep = ""), sep = ""))
fo <- as.formula(paste("class ~", paste(xvars, collapse=" +" )))
d <- maipofields

wrp <- pca_warper(d, xvars = xvars, yvar = "class")
plot(wrp)
round(wrp$pca$rotation[,1:3], digits = 2)
summary(wrp)

# Create a data frame with transformed data:
wd <- warp(d, warper = wrp)
summary(wd)

# Backtransform it, should be identical to d:
d2 <- unwarp(wd, warper = wrp)
all.equal(d, d2[,colnames(d)])
# Default tolerance works for this data set, but you may have to use
# e.g. tol = 10^(-6) for less well conditioned data sets and
# transformations.
}
\seealso{
\code{\link[=warp]{warp()}} for the forward transformation, and
\code{\link[=pca_warper]{pca_warper()}} for an example of a simple warper function based
on the principal components transformation.
}
