% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/warper-functions.R
\name{warp.data.frame}
\alias{warp.data.frame}
\alias{warp.warped_df}
\title{Transform data from feature space into a transformed space}
\usage{
\method{warp}{data.frame}(x, warper)

\method{warp}{warped_df}(x, warper)
}
\arguments{
\item{x}{The object to be transformed, normally a data frame.}

\item{warper}{A \code{warper} object as, for example, created by \code{\link{pca_warper}} for principal components transformations.}

\item{...}{Additional arguments to be passed to the \code{warp} method.}
}
\value{
A data frame with features in the transformed (e.g., PCA) space.
}
\description{
\code{warp} methods apply the transformation \eqn{T} to data
in feature space - the result is, for example, the principal components
representation of the data, if a \code{\link{pca_warper}} is used.
}
\section{Functions}{
\itemize{
\item \code{warp.warped_df}: Does nothing - \code{warped_df} has already been transformed
}}

\seealso{
\code{\link{unwarp}} for the inverse transformation, and
\code{\link{pca_warper}} for an example of a simple warper function based
on the principal components transformation.
}
